--- a/Project.toml
+++ b/Project.toml
@@ -13,7 +13,7 @@ SpecialFunctions = "276daf66-3868-5448-9aa4-cd146d93841b"
 
 [compat]
 BenchmarkTools = ">=1.3.2"
-Documenter = ">=0.27.24"
+Documenter = ">=1.0.0"
 LaTeXStrings = ">=1.3"
 Plots = ">=1.12.4"
 QuadGK = ">=2.4.2"