--- a/docs/Project.toml
+++ b/docs/Project.toml
@@ -6,6 +6,6 @@ BenchmarkTools = "6e4b80f9-dd63-53aa-95a3-0cdb28fa8baf"
 
 [compat]
 BenchmarkTools = "1.3.2"
-Documenter = "0.27.24"
+Documenter = ">=1.0.0"
 LaTeXStrings = "1.3"
 Plots = "1.12.4"